requirements.txt
pandas==1.5.3
numpy==1.26.4
pyyaml==6.0.1
pytest==7.4.4
